+++ src/Arrays.php
@@ -545,6 +545,7 @@ class Arrays implements ArrayAccess, Countable, IteratorAggregate
     public function sortBy(string $key, string $direction = 'ASC', int $sortFlags = SORT_REGULAR): self
     {
         $array  = $this->items;
+        $direction = mb_strtolower($direction);
         $result = [];
 
         if (count($array) <= 0) {
@@ -557,8 +558,8 @@ class Arrays implements ArrayAccess, Countable, IteratorAggregate
 
         if ($sortFlags === SORT_NATURAL) {
             natsort($helper);
-            ($direction === 'DESC') and $helper = array_reverse($helper);
-        } elseif ($direction === 'DESC') {
+            ($direction === 'desc') and $helper = array_reverse($helper);
+        } elseif ($direction === 'desc') {
             arsort($helper, $sortFlags);
         } else {
             asort($helper, $sortFlags);
@@ -571,6 +572,34 @@ class Arrays implements ArrayAccess, Countable, IteratorAggregate
         $this->items = $result;
 
         return $this;
+    }
+
+    /**
+     * Sorts a associative array by a certain key in descending order.
+     *
+     * @param  string $key    The name of the key.
+     * @param  int    $sortFlags A PHP sort method flags.
+     *                           https://www.php.net/manual/ru/function.sort.php
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function sortByDesc(string $key, int $sortFlags = SORT_REGULAR): self
+    {
+        return $this->sortBy($key, 'DESC', $sortFlags);
+    }
+
+    /**
+     * Sorts a associative array by a certain key in ascending order.
+     *
+     * @param  string $key    The name of the key.
+     * @param  int    $sortFlags A PHP sort method flags.
+     *                           https://www.php.net/manual/ru/function.sort.php
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function sortByAsc(string $key, int $sortFlags = SORT_REGULAR): self
+    {
+        return $this->sortBy($key, 'ASC', $sortFlags);
     }
 
     /**
@@ -1029,6 +1058,18 @@ class Arrays implements ArrayAccess, Countable, IteratorAggregate
     }
 
     /**
+     * Skip the first count items.
+     *
+     * @param  int  $count Count of first items to skip.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function skip(int $count): self
+    {
+        return $this->slice($count);
+    }
+
+    /**
      * Verifies that all elements pass the test of the given callback.
      *
      * @param Closure $callback Function with (value, key) parameters and returns TRUE/FALSE
@@ -1459,7 +1500,7 @@ class Arrays implements ArrayAccess, Countable, IteratorAggregate
     /**
      * Filters the array items by a given condition.
      *
-     * @param string $key      Key of the array or object to used for comparison.
+     * @param string $key      Key of the array for comparison.
      * @param string $operator Operator used for comparison.
      *                         operators: in, nin, lt, <, lte,
      *                                    >, gt, gte, >=, contains, ncontains
@@ -1562,6 +1603,240 @@ class Arrays implements ArrayAccess, Countable, IteratorAggregate
         );
 
         return $this;
+    }
+
+    /**
+     * Filters the array items by the given key value pair.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereIn(string $key, $value): self
+    {
+        return $this->where($key, 'in', $value);
+    }
+
+    /**
+     * Filters the array items by the given key value pair.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereNotIn(string $key, $value): self
+    {
+        return $this->where($key, 'nin', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is between the given values.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereBetween(string $key, $value): self
+    {
+        return $this->where($key, 'between', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is not between the given values.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereNotBetween(string $key, $value): self
+    {
+        return $this->where($key, 'nbetween', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is less the given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereLess(string $key, $value): self
+    {
+        return $this->where($key, 'lt', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is less or equal the given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereLessOrEqual(string $key, $value): self
+    {
+        return $this->where($key, 'lte', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is greater the given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereGreater(string $key, $value): self
+    {
+        return $this->where($key, 'gt', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is greater or equal the given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereGreaterOrEqual(string $key, $value): self
+    {
+        return $this->where($key, 'gte', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is contains given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereContains(string $key, $value): self
+    {
+        return $this->where($key, 'contains', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is not contains given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereNotContains(string $key, $value): self
+    {
+        return $this->where($key, 'ncontains', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is equal given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereEqual(string $key, $value): self
+    {
+        return $this->where($key, 'eq', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is not equal given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereNotEqual(string $key, $value): self
+    {
+        return $this->where($key, 'neq', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is starts with given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereStartsWith(string $key, $value): self
+    {
+        return $this->where($key, 'starts_with', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is ends with given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereEndsWith(string $key, $value): self
+    {
+        return $this->where($key, 'ends_with', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is newer given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereNewer(string $key, $value): self
+    {
+        return $this->where($key, 'newer', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is older given value.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereOlder(string $key, $value): self
+    {
+        return $this->where($key, 'older', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is matches to given regexp.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereRegexp(string $key, $value): self
+    {
+        return $this->where($key, 'regexp', $value);
+    }
+
+    /**
+     * Filters the array items by the given key is not matches to given regexp.
+     *
+     * @param string $key   Key of the array for comparison.
+     * @param mixed  $value Value used for comparison.
+     *
+     * @return self Returns instance of The Arrays class.
+     */
+    public function whereNotRegexp(string $key, $value): self
+    {
+        return $this->where($key, 'nregexp', $value);
     }
 
     /**